+++ src/users/users.controller.ts
@@ -15,7 +15,7 @@ import {
 import { CreateUserDto } from './dto/create-user.dto';
 import { UpdatePasswordDto } from './dto/update-user.dto';
 import { UserEntity } from './entity/userEntity';
-import { UsersService } from './user.service';
+import { UsersService } from './users.service';
 @UseInterceptors(ClassSerializerInterceptor)
 @Controller('user')
 export class UsersController {
